backend/requirements.txt
fastapi
uvicorn[standard]
langchain
openai
chromadb
python-dotenv
sqlalchemy
psycopg2-binary 